conflits resolve problems

+++ src/app/users/page.js
@@ -7,6 +7,7 @@ import Edit from "@/components/Edit";
 import Button from "@/components/UI/Button";
 import CardUser from "@/components/UI/card/user_card";
 import FormUser from "@/components/UI/form/user_form";
+import { Toaster } from "react-hot-toast";
 
 export default function Users() {
   const title = "users";
@@ -39,11 +40,16 @@ export default function Users() {
   }, [db_name]);
 
   const deleteData = async (id) => {
-    try {
-      await FetchApi({ url: `/api/${db_name}/${id}`, method: "DELETE" });
-      setDataList(dataList.filter((item) => item.id !== id));
-    } catch (error) {
-      console.error("Deletion error: ", error);
+    const confirmDelete = window.confirm(
+      "Êtes-vous sûr de vouloir supprimer cet élément ?"
+    );
+    if (confirmDelete) {
+      try {
+        await FetchApi({ url: `/api/${db_name}/${id}`, method: "DELETE" });
+        setDataList(dataList.filter((item) => item.id !== id));
+      } catch (error) {
+        console.error("Deletion error: ", error);
+      }
     }
   };
 
@@ -60,8 +66,7 @@ export default function Users() {
           dataList={dataList}
         />
       )}
-
-
+      <Toaster />
       <div className={styles.listContainer}>
         <h1 className={styles.listTitle}>{title}</h1>
         {dataList.map((data) => (
